horologium-api/requirements.txt
django==3.0.7
djangorestframework==3.11.0
django_html_sanitizer==0.1.5